Made usage of gsf optional
