+++ src/CMakeLists.txt
@@ -36,7 +36,7 @@ option(LIBIGL_WITH_XML               "Use XML"            OFF)
 #option(LIBIGL_EXPORT_TARGETS         "Export libigl CMake targets" ON)
 
 option(AUVLIB_WITH_GSF "Enable reading and processing gsf data" ON)
-option(AUVLIB_USE_LIBIGL_EIGEN "Use libigl's version of Eigen"  ON)
+option(AUVLIB_USE_LIBIGL_EIGEN "Use libigl's version of Eigen"  OFF)
 option(AUVLIB_USE_LIBIGL_GLFW "Use libigl's version of glfw"    OFF)
 option(AUVLIB_USE_PYTHON3 "Use python3 instead of python2"      OFF)
 
@@ -88,7 +88,9 @@ link_directories(${Boost_LIBRARY_DIRS})
 #include_directories(eigen_cereal/include)
 
 include(cmake/glad.cmake)
-add_subdirectory(libgsf)
+if(AUVLIB_WITH_GSF)
+  add_subdirectory(libgsf)
+endif()
 add_subdirectory(libxtf)
 add_subdirectory(eigen_cereal)
 add_subdirectory(data_tools)
