+++ src/data_tools/CMakeLists.txt
@@ -11,7 +11,10 @@ add_library(benchmark src/benchmark.cpp)
 
 add_library(navi_data src/navi_data.cpp)
 
-add_library(gsf_data src/gsf_data.cpp)
+if(AUVLIB_WITH_GSF)
+  add_library(gsf_data src/gsf_data.cpp)
+  add_executable(test_submap_tracks src/test_submap_tracks.cpp)
+endif()
 
 add_library(csv_data src/csv_data.cpp)
 
@@ -24,8 +27,6 @@ add_library(all_data src/all_data.cpp)
 add_library(xyz_data src/xyz_data.cpp)
 
 add_executable(test_xtf src/test_xtf.cpp)
-
-add_executable(test_submap_tracks src/test_submap_tracks.cpp)
 
 add_executable(test_all src/test_all.cpp)
 
@@ -57,10 +58,12 @@ target_include_directories(navi_data PUBLIC
     $<INSTALL_INTERFACE:include>
     PRIVATE src)
 
-target_include_directories(gsf_data PUBLIC
-    $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
-    $<INSTALL_INTERFACE:include>
-    PRIVATE src)
+if(AUVLIB_WITH_GSF)
+  target_include_directories(gsf_data PUBLIC
+      $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
+      $<INSTALL_INTERFACE:include>
+      PRIVATE src)
+endif()
 
 target_include_directories(csv_data PUBLIC
     $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
@@ -89,8 +92,6 @@ target_include_directories(xyz_data PUBLIC
 
 target_link_libraries(test_xtf xtf_data std_data navi_data ${OpenCV_LIBS} ${Boost_LIBRARIES})
 
-target_link_libraries(test_submap_tracks gsf_data std_data navi_data ${OpenCV_LIBS} ${Boost_LIBRARIES})
-
 target_link_libraries(test_all all_data ${OpenCV_LIBS} ${Boost_LIBRARIES})
 
 # Link the libraries
@@ -102,9 +103,12 @@ target_link_libraries(benchmark PUBLIC eigen_cereal std_data ${OpenCV_LIBS})
 
 target_link_libraries(navi_data PUBLIC eigen_cereal data_transforms) # ${PCL_LIBRARIES})
 
-target_link_libraries(gsf_data PUBLIC gsf std_data navi_data lat_long_utm)
+target_link_libraries(csv_data PUBLIC std_data navi_data xtf_data)
 
-target_link_libraries(csv_data PUBLIC std_data navi_data gsf_data xtf_data)
+if(AUVLIB_WITH_GSF)
+  target_link_libraries(gsf_data PUBLIC gsf std_data navi_data lat_long_utm csv_data)
+  target_link_libraries(test_submap_tracks gsf_data std_data navi_data ${OpenCV_LIBS} ${Boost_LIBRARIES})
+endif()
 
 target_link_libraries(xtf_data PUBLIC std_data navi_data xtf_reader lat_long_utm ${OpenCV_LIBS})
 
@@ -112,8 +116,14 @@ target_link_libraries(all_data navi_data lat_long_utm csv_data ${OpenCV_LIBS})
 
 target_link_libraries(xyz_data std_data ${Boost_LIBRARIES})
 
+set(AUVLIB_DATA_TOOLS_LIBS data_transforms submaps std_data benchmark navi_data csv_data xtf_data all_data xyz_data lat_long_utm)
+
+if(AUVLIB_WITH_GSF)
+  set(AUVLIB_DATA_TOOLS_LIBS ${AUVLIB_DATA_TOOLS_LIBS} gsf_data)
+endif()
+
 # 'make install' to the correct locations (provided by GNUInstallDirs).
-install(TARGETS data_transforms submaps std_data benchmark navi_data gsf_data csv_data xtf_data all_data xyz_data lat_long_utm EXPORT DataToolsConfig
+install(TARGETS ${AUVLIB_DATA_TOOLS_LIBS} EXPORT DataToolsConfig
     ARCHIVE  DESTINATION ${CMAKE_INSTALL_LIBDIR}
     LIBRARY  DESTINATION ${CMAKE_INSTALL_LIBDIR}
     RUNTIME  DESTINATION ${CMAKE_INSTALL_BINDIR})  # This is for Windows
@@ -125,4 +135,4 @@ install(DIRECTORY include/ DESTINATION ${CMAKE_INSTALL_INCLUDEDIR})
 install(EXPORT DataToolsConfig DESTINATION share/DataTools/cmake)
 
 # This makes the project importable from the build directory
-export(TARGETS data_transforms submaps std_data benchmark navi_data gsf_data csv_data xtf_data all_data xyz_data lat_long_utm FILE DataToolsConfig.cmake)
+export(TARGETS ${AUVLIB_DATA_TOOLS_LIBS} FILE DataToolsConfig.cmake)
