+++ src/pydata_tools/CMakeLists.txt
@@ -4,7 +4,9 @@ message("Python dir: " ${PYTHON_INSTALL_PATH})
 
 #pybind11_add_module(pystd_data src/pystd_data.cpp)
 add_library(pystd_data MODULE src/pystd_data.cpp)
-add_library(pygsf_data MODULE src/pygsf_data.cpp)
+if(AUVLIB_WITH_GSF)
+  add_library(pygsf_data MODULE src/pygsf_data.cpp)
+endif()
 add_library(pyxtf_data MODULE src/pyxtf_data.cpp)
 add_library(pycsv_data MODULE src/pycsv_data.cpp)
 add_library(pyall_data MODULE src/pyall_data.cpp)
@@ -16,10 +18,12 @@ target_include_directories(pystd_data PUBLIC
     $<INSTALL_INTERFACE:include>
     PRIVATE src)
 
-target_include_directories(pygsf_data PUBLIC
-    $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
-    $<INSTALL_INTERFACE:include>
-    PRIVATE src)
+if(AUVLIB_WITH_GSF)
+  target_include_directories(pygsf_data PUBLIC
+      $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
+      $<INSTALL_INTERFACE:include>
+      PRIVATE src)
+endif()
 
 target_include_directories(pyxtf_data PUBLIC
     $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
@@ -53,10 +57,12 @@ set_target_properties(pystd_data PROPERTIES PREFIX "${PYTHON_MODULE_PREFIX}"
                                             OUTPUT_NAME "std_data"
                                             SUFFIX "${PYTHON_MODULE_EXTENSION}")
 
-target_link_libraries(pygsf_data PRIVATE std_data gsf_data eigen_cereal ${BOOST_LIBRARIES} pybind11::module)
-set_target_properties(pygsf_data PROPERTIES PREFIX "${PYTHON_MODULE_PREFIX}"
-                                            OUTPUT_NAME "gsf_data"
-                                            SUFFIX "${PYTHON_MODULE_EXTENSION}")
+if(AUVLIB_WITH_GSF)
+  target_link_libraries(pygsf_data PRIVATE std_data gsf_data eigen_cereal ${BOOST_LIBRARIES} pybind11::module)
+  set_target_properties(pygsf_data PROPERTIES PREFIX "${PYTHON_MODULE_PREFIX}"
+                                              OUTPUT_NAME "gsf_data"
+                                              SUFFIX "${PYTHON_MODULE_EXTENSION}")
+endif()
 
 target_link_libraries(pyxtf_data PRIVATE std_data xtf_data eigen_cereal ${BOOST_LIBRARIES} pybind11::module)
 set_target_properties(pyxtf_data PROPERTIES PREFIX "${PYTHON_MODULE_PREFIX}"
@@ -83,7 +89,12 @@ set_target_properties(pybenchmark PROPERTIES PREFIX "${PYTHON_MODULE_PREFIX}"
                                              OUTPUT_NAME "benchmark"
                                              SUFFIX "${PYTHON_MODULE_EXTENSION}")
 
-install(TARGETS pystd_data pygsf_data pyxtf_data pycsv_data pyall_data pyxyz_data pybenchmark
+set(AUVLIB_PYDATA_TOOLS_LIBS pystd_data pyxtf_data pycsv_data pyall_data pyxyz_data pybenchmark)
+if(AUVLIB_WITH_GSF)
+  set(AUVLIB_PYDATA_TOOLS_LIBS ${AUVLIB_PYDATA_TOOLS_LIBS} pygsf_data)
+endif()
+
+install(TARGETS 
     LIBRARY  DESTINATION ${CMAKE_INSTALL_LIBDIR}/auvlib/data_tools
 )  # This is for Windows
